auth/user_auth/requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
contourpy==1.0.6
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.1.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
djoser==2.2.0
efficientnet==1.0.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.2
Flask==2.2.3
flatbuffers==23.5.8
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.49.1
h5py==3.7.0
huggingface-hub==0.16.4
idna==3.4
image-classifiers==1.0.0
imageio==2.28.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.9
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
keras==2.12.0
Keras-Applications==1.0.8
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.1.0
mysqlclient==2.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
patchify==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
plumbum==1.8.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.23.0
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-build==0.16.7
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.1
segmentation-models==1.0.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tokenizers==0.13.3
torch==1.12.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.31.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.4
wrapt==1.14.1
